[ACD-8977] Take acad exe name from user config if possible (#64)
Fix [ACD-8977]: Take acad exe name from user config if possible.

The old logic is to pick up the name of acad from launch.json. But it's possible that there are multiple configurations for user to pick up one. In this case, we shall take the one that the user selected.