Sign the extension before publishing (#84)
This PR signs the extension before publishing.

Note that the official sign tool by Microsoft, VsixSignTool, is not cross-platform. So use OpenOpcSignTool , an open-source tool that based on .Net Core, instead.

The tool is installed in the updated docker image, autodesk-docker.art-bobcat.autodesk.com:10873/autolisp-ext/build:2.0. And all signing credentials are stored on artifactory.