Add Whitesource to Jenkins and use yarn and some updates. (#129)
ACD-11432: add whitesource check.
In order for it to work, we need to use yarn package. (npm will report false alarm that is hard to fix)
I also removed dependency of vscode-i18n as it is not used.
vscode-test framework has been changed, so I updated it. Later, I will make it work.
lodash* package is already latest but it still has security issue. As it is only used by development enviroment, I exclude it for now.
Besides, I add code to report build failure.