fix bug ACD-11801 (#165)
This commit is to fix the bug ACD-11801.

### Abstract
The files list in the project file is a type of long list, and it is controlled by the setting "long list format style".
kWideStyleSingleCol means this:
![image](https://media.git.autodesk.com/user/4246/files/fd9b2b00-a103-11ea-9aad-f76b35c8d126)

kFitToMargin means this:
![image](https://media.git.autodesk.com/user/4246/files/07bd2980-a104-11ea-8721-30cc8a7de03b)

This two format is not the same as the legacy VLIDE when format the project file. 
So I introduce the third format: `kSingleColumn`, this is real single colimn with Narrow style

### Note to TD
Please retest the bug
And test the format for "`long list format style`"
Then test the case if the file list is less than 5 items, and larger than 5 items.