improve summary node tooltip for search tree (#181)
If search results exceeds the max number, we will add a message in the summary. But it's too long to show the whole string, user have to extend the Find & Replace box to see the whole summary. It can't be displayed in two lines, so I add the summary in the tooltip.