ACD-12078 Pop Error: stdout maxBuffer exceeded or "no results found" when too many search result in one file (#188)
For "stdout maxBuffer exceeded", this is because the search results exceed the max buffer allowed on stdout of child process (search tool rg.exe is running in a child process by execa).

child_process.spawn may solve this issue, but I think we can just catch the error and show "Limit of search results exceeded. Refine your search to narrow down the results." for now.

For "no results found" , this is because the search results exceed the max results number (10000), so the results are not displayed in the search tree. But it should not show "No results found", we can show
"Limit of search results exceeded. Refine your search to narrow down the results."