Fix CI and clean some legal text in source files (#203)
Readme processing was broken during packaging
due to the fact that it couldn't figure out how to
replace the link in Notice.md. We now specify the
repo to help with this.

Some law text in ts file are not correct, so I clean up all the law text in ts files after reviewed the whole project.