Setup Ci (#42)
We use Github Actions to build every PR and we also create a release for every push into main. The release contains snapshot of the code + the vsix. The vsix is versioned as prerelease. We use npm version to create a new prerelease version number during the build in package.json (we don't commit this change).