Doc manager update (#49)
Objective
This is an update to the DocumentManager to persist/recycle more data. It was decided during the DefinitionProvider PR #45 that something had to be done about about the excessively expensive ReadonlyDocument.atomsForest being regenerated more often than it needed to be.

Abstraction/Implementation
This change involves a major emphasis on tracking pairs ReadonlyDocument's & vscode.TextDocuments and performing very cheap checks to detect when the ReadonlyDocument is out of date. It uses a set of flags to specify the various contextual applications. This will inevitably make a ReadonlyDocument qualify for up to 4 queries

Currently Active Document
Opened
Project
Workspace
However, the expensive parts should now be completely recycled and very cheap checks are in place to avoid the excessive parts of that potential duplication. From what I can tell, nothing should ever go through all 4 in expensive ways because any one would have passed the cheap pre-check, had the expensive work done and never moved on to the others.

The original version of this did a lot better job of purging dead data, but I believe this will at least make it inaccessible. This is notable for file renames and deletions performed directly in a OS file explorer. In theory, this is removing the flags that qualify them for various queries, but not necessarily removing the reference entirely across all possible situations. I believe the actions performed in the vscode workspace itself would do proper data purging purging though.

This also removed the PRJ embedded ReadonlyDocument property. Everything is demand queried from the DocumentManager when applicable, but persisted so it can efficiently serve it back next time. The cheap pre-check added in the DefinitionProvider and the fact that ReadonlyDocument creation is also quite cheap should adequately compensate for this. Note that PRJ's being opened/closed is probably the biggest issue with persisted dead data. Since the the ProjectTree's Instance() is always used to produce the "single source of truth" of fspath keys this should not actually produce erroneous PRJ references, but it has no mechanism for getting rid of the old ones. I'd like to maintain a maximum separation between the Document Manager and the Project Manager if we can.

Removed the attempts to persist data across workspace changes. From what I can tell, this was an exercise of futility. I tried multiple tests to verify it was working and kept eventually winding up with an empty _cached property. For example the current arrangement of the initialize() method that setups the cache isn't even clearing it anymore and all references are still disappearing after any action that changes the workspace.

Honestly the documents.ts is vastly more clustered than it used to be since it is now handling various edge cases and was massively changed. So, I may be easier to perform your first review without the Adds/Deletes turned on.

Tests Passed
I kind of just described them above by identifying the potential known issues. I did run this with the F12 on a Workspace only, a PRJ only and a Workspace/PRJ hybrid.

Screenshots
Not Applicable