make correct package for vsix file in Github CI (#63)
##### Objective
Github CI didn't package the visx file correctly, it needs to call the pack.py to copy some utility files to correct location.

##### Abstractions
In the pack.py it runs the `npm install`, and copy some utilities files for example the search tool exe files to the correct locations, then `gulp package `can make an expected vsix file for us. So it means that the several pre-release packages we made in CI are not correct. (I will do house keeping latter).

I also make the gulp to install it locally instead of globally, because CI machine to complains that it has no access to some global path.

##### Tests performed
Checked the PR preflight messages that there is no warning to run the pack.py.

##### Screen shot
Not applicable for this PR