Deprecate to use the launch.json to debug autolisp files (#55)
#### Objectives
In this commit I trid to remove the support for the launchjson.

#### Abstraction
Now the debug settings (attach and launch for acad installation path) is in the extension "settings" page, which is a global settings. So it has no need to use the launch.json file. In fact the file has been deprecated in the version 1.3.0, and it only reads the settings (can not set the settings).

In this version I tried to remove the reading setting support for launch.json, it has some issues to debug the lisp after upgrade the vscode. Some customers reported that after updating vscode, it may have issue to attach debugging autocad. 

#### tests performed
tried to open a folder with the legacy launch.json, and do following settings:
 1. attach debug
 2. launch debug
 Then do some changes in the launch.json file, and do above tests again

#### Screen Shot
No new behavior introduced, not applicable