refine CHANGELOG and fix online help bug (#67)
##### Objective
In this commit I want to fix the online help bug, because the latest AutoCAD has not been released, so the ACAD2022 help can not get. And I also refined the changelog text.

##### Abstractions
Our tester complains these two issues to us that the "online help" is broken at this point that the new AutoCAD 2022 has not been released. And the change log text is also not correct. So in this commit I aimed to resolve those two issues.

##### Tests performed
- to test the "Online help" can work well for built in keywords

##### Screen shot
Not applicable for this PR