remove the duplicate complete item words ACD-17606 (#77)
##### Objective
I want to fix the issue that when user input `vlisp` in the vscode auto-completion list it has some duplicates items.
![image](https://user-images.githubusercontent.com/68938334/106558078-02588000-64d8-11eb-9a40-11976079488b.png)

And these `vlisp- `prefix functions should not list in MacOS because they are not supported in that platform. I also plan to fix it.

##### Abstractions
This is a bug found by the tester as ACD-17606. This is only to change in the extension side, so I plan to fix it.

I also remove the un-needed Jienken file which is only used in the Autodesk internal pipeline, but not the pipeline is retired. So remove it here.

##### Tests performed
 - rebuild the extension and try to input keys like `vlisp-`, then verify the auto-completion items are expected.

##### Screen shot